requirements.txt
Django==5.2
Pillow==12.2.0
django-admin-sortable2==2.3.1
